chore: Add file and install node modules if they don't exist
This commit adds a new Makefile function called `copyFileIfNotExist` that copies a file from one location to another if it doesn't already exist. It also adds another function called `installNodeModulesIfNotExist` that installs the node modules if they are not already present.

The purpose of these changes is to ensure that the required files and node modules are present before running the development environment using the `make dev` command.